agent_swarm/requirements.txt
langchain
langchain_openai
langgraph
requests
gitpython
slack_sdk
python-dotenv